+++ src/XrmContext/AssemblyInfo.fs
@@ -5,10 +5,10 @@ open System.Reflection
 [<assembly: AssemblyProductAttribute("XrmContext")>]
 [<assembly: AssemblyDescriptionAttribute("Tool to generate early-bound .NET framework classes and enumerations for MS CRM Dynamics server-side coding.")>]
 [<assembly: AssemblyCompanyAttribute("Delegate A/S")>]
-[<assembly: AssemblyCopyrightAttribute("Copyright (c) Delegate A/S 2015")>]
-[<assembly: AssemblyVersionAttribute("1.1.2")>]
-[<assembly: AssemblyFileVersionAttribute("1.1.2")>]
+[<assembly: AssemblyCopyrightAttribute("Copyright (c) Delegate A/S 2016")>]
+[<assembly: AssemblyVersionAttribute("1.2.0")>]
+[<assembly: AssemblyFileVersionAttribute("1.2.0")>]
 do ()
 
 module internal AssemblyVersionInformation =
-    let [<Literal>] Version = "1.1.2"
+    let [<Literal>] Version = "1.2.0"