+++ src/XrmContext/Utility.fs
@@ -8,8 +8,15 @@ open System.Text.RegularExpressions
 module Utility =
 
   let (|>>) x g = g x; x
-
+  let (?>>) m f = Option.bind f m
+  let (?|>) m f = Option.map f m
   let (|?) = defaultArg
+
+  let parseInt str =
+    let mutable intvalue = 0
+    if System.Int32.TryParse(str, &intvalue) then Some(intvalue)
+    else None
+
   let (|StartsWithNumber|) (str:string) = str.Length > 0 && str.[0] >= '0' && str.[0] <= '9'
   let (|StartsWith|_|) needle (haystack : string) = if haystack.StartsWith(needle) then Some() else None
 
@@ -33,4 +40,14 @@ module Utility =
     let obj = (new DataContractJsonSerializer(typeof<'t>)).ReadObject(ms) 
     obj :?> 't
 
-  
+  let parseVersion (str:string) =
+    let vArr = str.Split('.')
+    let getIdx idx = Array.tryItem idx vArr ?>> parseInt |? 0
+    (getIdx 0, getIdx 1, getIdx 2, getIdx 3)
+
+  let checkVersion (a1,b1,c1,d1) (a2,b2,c2,d2) =
+    ([a2;b2;c2;d2], [a1;b1;c1;d1])
+    ||> List.map2 (-)
+    |> List.tryFind ((<>) 0)
+    ?|> fun x -> x > 0
+    |? true